requirements.txt
Flask
Flask-Migrate
Flask-Login
psycopg2-binary
python-dotenv
python-slugify
SQLAlchemy
text-unidecode
typing_extensions
Werkzeug
WTForms
gunicorn
